requirements.txt
django
django-allauth
django-cors-headers
django-environ
django-filter
djangorestframework
djangorestframework-camel-case
djangorestframework_simplejwt
dj-rest-auth
gunicorn
pillow
psycopg2-binary
requests
uvicorn
whitenoise[brotli]
cloudinary